+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Removed ICS precompile
 
 ## Fixed
+- Fix tokenfactory mintTo to check blocked address before minting ([#258](https://github.com/KiiChain/kiichain/issues/258))
 - Return error instead of nil in RemoveExcessFeeds to properly propagate storage errors
 - Fixed oracle module ConsensusVersion constant not being used ([#256](https://github.com/KiiChain/kiichain/issues/256))
 - Fix oracle weighted median threshold to require >50% instead of >=50% for majority